Hook up stateful logic to list sidebar view
